+++ src/main.ts
@@ -1,6 +1,13 @@
 import DiscordManager from "src/discord/DiscordManager";
 import LocalImageModal from "src/LocalImageModal";
-import { Notice, Plugin, TFile, Workspace } from "obsidian";
+import {
+	Editor,
+	FileSystemAdapter,
+	Notice,
+	Plugin,
+	TFile,
+	Workspace,
+} from "obsidian";
 import SettingsTab, {
 	DEFAULT_VALUES,
 	INITIAL_SETTINGS,
@@ -8,6 +15,7 @@ import SettingsTab, {
 	PartialSettings,
 } from "src/Settings";
 import DiscordHelper from "./discord/DiscordHelper";
+import { DiscordEmbedParams } from "./discord/types";
 
 export default class DiscordSharePlugin extends Plugin {
 	settings: ISettingsOptions;
@@ -50,7 +58,17 @@ export default class DiscordSharePlugin extends Plugin {
 					return !!file && !!discordWebhookURLSet;
 				}
 				if (file instanceof TFile) {
-					this.discordManager.shareEmbed(file);
+					const params =
+						this.discordHelper.buildDiscordEmbedParamsFromFile(
+							file
+						);
+					if (!params) {
+						new Notice(
+							`Failed to build Discord embed params from file ${file.name}.`
+						);
+						return;
+					}
+					this.discordManager.shareEmbed(params);
 				} else {
 					new Notice("No active file found.");
 				}
@@ -58,22 +76,43 @@ export default class DiscordSharePlugin extends Plugin {
 		});
 
 		this.addCommand({
-			id: "discord:share-title",
-			name: "Share Note Title to Discord",
-			checkCallback: (checking) => {
+			id: "discord:share-selection",
+			name: "Share Selection to Discord",
+			editorCallback: async (editor: Editor) => {
 				const discordWebhookURLSet =
 					this.getSettingValue("discordWebhookURL");
-				const file = this.workspace.getActiveFile();
-				if (checking) {
-					return !!file && !!discordWebhookURLSet;
-				}
-				if (file instanceof TFile) {
-					this.discordManager.shareFileTitle(file.basename);
-				} else {
-					new Notice("No active file found.");
-				}
+				if (!discordWebhookURLSet || !editor.somethingSelected())
+					return;
+				const selection = editor.getSelection().trim();
+				const params: Partial<DiscordEmbedParams> = {
+					description: selection,
+				};
+				this.discordManager.shareEmbed(params);
 			},
 		});
+
+		this.registerEvent(
+			this.app.workspace.on("editor-menu", (menu, editor) => {
+				const discordWebhookURLSet =
+					this.getSettingValue("discordWebhookURL");
+				if (!(this.app.vault.adapter instanceof FileSystemAdapter))
+					return;
+				if (!discordWebhookURLSet || !editor.somethingSelected())
+					return;
+				const selection = editor.getSelection().trim();
+
+				menu.addItem((item) => {
+					item.setTitle("Share selection to Discord").onClick(
+						async () => {
+							const params: Partial<DiscordEmbedParams> = {
+								description: selection,
+							};
+							this.discordManager.shareEmbed(params);
+						}
+					);
+				});
+			})
+		);
 	}
 
 	onunload() {}